+++ app/app/tests/abstract/mock_view.py
@@ -1,7 +1,7 @@
 
 from django.contrib.auth.models import User
 
-from access.middleware.auth import Tenancy
+from access.middleware.request import Tenancy
 from access.models import Organization
 
 from access.models import Organization
