+++ app/core/tests/abstract/test_ticket_serializer.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.test import TestCase
 
-from access.middleware.auth import Tenancy
+from access.middleware.request import Tenancy
 from access.models import Organization, Team, TeamUsers, Permission
 
 from rest_framework.exceptions import ValidationError
