+++ app/itam/tests/functional/device_operating_system/test_device_operating_system_serializer.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 from rest_framework.exceptions import ValidationError
 
-from access.middleware.auth import Tenancy
+from access.middleware.request import Tenancy
 from access.models import Organization
 
 from itam.serializers.device_operating_system import Device, DeviceOperatingSystem, DeviceOperatingSystemModelSerializer
