+++ app/itam/tests/functional/device_software/test_device_software_serializer.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 from rest_framework.exceptions import ValidationError
 
-from access.middleware.auth import Tenancy
+from access.middleware.request import Tenancy
 from access.models import Organization
 
 from itam.serializers.device_software import Device, DeviceSoftware, DeviceSoftwareModelSerializer
